Restructure test dir
